+++ src/main/java/amidst/mojangapi/file/service/AmidstBackupService.java
@@ -1,4 +1,4 @@
-package amidst.mojangapi.file;
+package amidst.mojangapi.file.service;
 
 import java.io.File;
 import java.io.IOException;
