+++ src/main/java/amidst/mojangapi/file/service/FilenameService.java
@@ -1,4 +1,4 @@
-package amidst.mojangapi.file;
+package amidst.mojangapi.file.service;
 
 import java.io.File;
