+++ src/main/java/amidst/mojangapi/world/WorldBuilder.java
@@ -4,9 +4,9 @@ import java.io.IOException;
 
 import amidst.documentation.Immutable;
 import amidst.mojangapi.file.MojangApiParsingException;
-import amidst.mojangapi.file.SaveDirectoryService;
 import amidst.mojangapi.file.directory.SaveDirectory;
 import amidst.mojangapi.file.nbt.LevelDatNbt;
+import amidst.mojangapi.file.service.SaveDirectoryService;
 import amidst.mojangapi.minecraftinterface.MinecraftInterface;
 import amidst.mojangapi.minecraftinterface.MinecraftInterfaceException;
 import amidst.mojangapi.minecraftinterface.RecognisedVersion;
