+++ src/main/java/amidst/mojangapi/world/player/WorldPlayerType.java
@@ -7,10 +7,10 @@ import java.util.List;
 
 import amidst.documentation.Immutable;
 import amidst.documentation.NotNull;
-import amidst.mojangapi.file.SaveDirectoryService;
 import amidst.mojangapi.file.directory.SaveDirectory;
 import amidst.mojangapi.file.nbt.LevelDatNbt;
 import amidst.mojangapi.file.nbt.player.PlayerNbt;
+import amidst.mojangapi.file.service.SaveDirectoryService;
 
 @Immutable
 public enum WorldPlayerType {
